Move directory so it's easier to migrate to TypeScripts
