+++ .eslintrc.json
@@ -3,9 +3,7 @@
         "browser": true,
         "es2021": true
     },
-    "extends": "airbnb-base",
-    "overrides": [
-    ],
+    "extends": ["airbnb-base"],
     "parserOptions": {
         "ecmaVersion": "latest",
         "sourceType": "module"
@@ -35,5 +33,30 @@
                 "ignoreImports": true
             }
         ]
-    }
+    },
+    "overrides": [
+        {
+            "files": ["lib/**/*.ts"],
+            "extends": [
+              "airbnb-typescript/base",
+              "plugin:@typescript-eslint/eslint-recommended",
+              "plugin:@typescript-eslint/recommended-type-checked"
+            ],
+            "parser": "@typescript-eslint/parser",
+            "parserOptions": {
+              "project": "./tsconfig.json"
+            },
+            "rules": {
+              "@typescript-eslint/explicit-module-boundary-types": "off",
+              "consistent-return": "off",
+              "import/prefer-default-export": "off",
+              "no-await-in-loop": "off",
+              "no-bitwise": "off",
+              "no-constant-condition": "off",
+              "no-continue": "off",
+              "no-param-reassign": "off",
+              "no-restricted-syntax": "off"
+            }
+        }
+    ]
 }
