+++ .gitignore
@@ -5,7 +5,3 @@ elprimobot-*
 
 # packaged sources in es5
 es5-packaged
-
-# logging
-exceptions.log
-rejections.log
