+++ README.md
@@ -6,7 +6,6 @@ It's a bot to help you manage Discord communities, especially those with program
 - Send the daily Leetcode challenge to a specific channel
 - Send the weekly Leetcode challenge to a specific channel
 - Daily and Weekly stats about the engagement of the community
-- Play rock, paper, scissors and more!
 
 ## Development
