+++ lib/api/leetcode/queries.js
