+++ lib/commands.js
