+++ lib/logging.js
