+++ package.json
@@ -15,6 +15,9 @@
   "bin": {
     "elprimobot": "es5-packaged/app.js"
   },
+  "prettier": {
+    "singleQuote": true
+  },
   "author": "Karl Jahn",
   "license": "MIT",
   "dependencies": {
@@ -33,8 +36,13 @@
     "@babel/cli": "^7.19.3",
     "@babel/core": "^7.19.3",
     "@babel/preset-env": "^7.19.3",
+    "@types/express": "^4.17.21",
+    "@types/node": "^20.11.26",
+    "@typescript-eslint/eslint-plugin": "^7.2.0",
+    "@typescript-eslint/parser": "^7.2.0",
     "eslint": "^8.26.0",
     "eslint-config-airbnb-base": "^15.0.0",
+    "eslint-config-airbnb-typescript": "^18.0.0",
     "eslint-plugin-import": "^2.26.0",
     "pkg": "^5.8.0"
   }
